requirements.txt
scikit_image==0.17.2
numpy==1.19.2
imutils==0.5.4
pandas==1.1.3
tensorflow==2.4.1
requests==2.24.0
scipy==1.5.2
opencv_python==4.5.1.48
matplotlib==3.3.2
Pillow==7.2.0
natsort==7.1.1

